Move embed logic to new crate and fix argv
