+++ src/interpreter.rs
@@ -1,5 +1,5 @@
 use crate::vm::{StackEntry, VMFnPrototype};
-use crate::{compiler, lexer, native, parser, state, stmt, value, vm};
+use crate::{compiler, embed, lexer, native, parser, state, stmt, value, vm};
 use std::collections::{HashMap, HashSet};
 use std::panic;
 use std::rc::Rc;
@@ -87,7 +87,7 @@ fn setup_global_interpreter() {
         );
         my_vm.builtins.insert(
             "process".to_string(),
-            value::Value::Native(native::Process::build()),
+            value::Value::Native(native::Process::build(embed::is_embedded())),
         );
         unsafe {
             GLOBAL_INTERPRETER = Some(Interpreter {
